feat: only show dahsboard if user is viewing own profile
